model/requirements.txt
google-generativeai==0.8.5
opencv-python==4.12.0.88
numpy==2.2.6
Pillow==11.3.0
pytesseract==0.3.13
dotenv==1.1.1
streamlit==1.49.1
torch==2.8.0
torchvision==0.23.0
scikit-learn==1.7.1
piexif==1.1.3
transformers==4.56.0
einops==0.8.1
